% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_outcome_for_approach.R
\name{adjust_outcome_for_approach}
\alias{adjust_outcome_for_approach}
\title{Adjust pitch outcome predictions for batter approach}
\usage{
adjust_outcome_for_approach(pred_outcome_pitch, approach, approach_model)
}
\arguments{
\item{pred_outcome_pitch}{a table with columns \code{prob_contact}, \code{prob_fair} and \code{pred_hit}}

\item{approach}{a table with columns \code{strikes_bat_speed} and \code{strikes_swing_length}}

\item{approach_model}{a list of fitted GLM models from \code{\link{fit_approach_model}}}
}
\value{
a table with approach-adjusted \code{prob_contact}, \code{prob_fair} and \code{pred_hit}
}
\description{
This function takes a table of n pitches (expressed as outcome predictions for \code{prob_contact},
\code{prob_fair} and \code{pred_hit}) and a table of m approaches (expressed as \code{strikes_bat_speed} and
\code{strikes_swing_length}) and returns a table of n * m predictions, one prediction for each pitch
for each approach.
}
