% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pred_outcome_pa.R
\name{calculate_pred_outcome_pa}
\alias{calculate_pred_outcome_pa}
\title{Calculate predicted plate appearance outcomes from predicted pitch outcomes}
\usage{
calculate_pred_outcome_pa(
  pred_outcome_by_count,
  linear_weight,
  nonterminal_prob_threshold = 1e-06
)
}
\arguments{
\item{pred_outcome_by_count}{a table with columns \code{balls}, \code{strikes}, \code{mean_prob_hbp},
\code{mean_prob_ball}, \code{mean_prob_strike}, \code{mean_prob_foul}, \code{mean_prob_fair} and \code{mean_pred_hit}}

\item{linear_weight}{a table with columns \code{event} and \code{linear_weight}}

\item{nonterminal_prob_threshold}{this numeric threshold determines how long we run the Markov
chain to calculate the terminal state probabilities from each starting state. We increase the
number of steps in the Markov chain until the probabilities for non-terminal states are below
\code{nonterminal_prob_threshold} (default 0.0001\%).}
}
\value{
a table with column \code{runs} reflecting the run value of a plate apperance with
the specificed pitch outcome distribution by count
}
\description{
In this function, we take a marginalized prediction across pitches within each count to get the
distribution of outcomes in each count and calculate the expected outcome of a plate appearance
assuming a Markov model from count to count. The input needs to have one row for each count,
but counts can be repeated if the the input table is grouped and each count appears exactly once
within each group (allowing you to simultaneously calculate the run value for plate appearances
under mutiple different assumptions).
}
