% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_approach.R
\name{evaluate_approach}
\alias{evaluate_approach}
\title{Evaluate an approach on the run scale}
\usage{
evaluate_approach(approach, pred_outcome_pitch, causal_model, linear_weight)
}
\arguments{
\item{approach}{a table with columns \code{strikes_bat_speed} and \code{strikes_swing_length}}

\item{pred_outcome_pitch}{a table with columns \code{prob_contact}, \code{prob_fair} and \code{pred_hit}}

\item{causal_model}{a list of fitted GLM models from \code{\link{fit_approach_model}}}

\item{linear_weight}{a table with columns \code{event} and \code{linear_weight}}
}
\value{
a table with columns \code{strikes_bat_speed}, \code{strikes_swing_length} and \code{runs}
}
\description{
This function takes some predicted pitch outcomes; adjusts them for approach; aggregates the
predictions by count; and rolls up the predictions into a run value assuming a Markov model for
the progression of a plate apperance. Approach is specified as a dataframe with a
\code{strikes_swing_length} column and a \code{strikes_bat_speed} column. This function can handle
multiple approaches (multiple rows in the approach table), but it can run into memory issues if
you try to pass too many (more than a hundred).
}
