% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_approach_model.R
\name{fit_approach_model}
\alias{fit_approach_model}
\title{Fit approach model}
\usage{
fit_approach_model(data_approach)
}
\arguments{
\item{data_approach}{a dataframe with at least the following columns: \code{strikes},
\code{strikes_bat_speed}, \code{strikes_swing_length}, \code{prob_contact}, \code{is_contact}, \code{prob_fair},
\code{is_fair}, \code{pred_hit} and \code{hit_pred}.}
}
\value{
a list with the following elements:
\itemize{
\item{"fit_contact"}{a fitted glm model for contact probability}
\item{"fit_fair"}{a fitted glm model for fair probability}
\item{"fit_hit"}{a fitted lm model for expected hit outcome}
}
}
\description{
This model estimates the effect of approach (modulating bat speed and swing length according to
count) on contact probability, fair probability and expected hit outcome.
}
