% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_intention_model_summary.R
\name{get_intention_model_summary}
\alias{get_intention_model_summary}
\title{Get intention model summary}
\usage{
get_intention_model_summary(
  intention_model_bat_speed,
  intention_model_swing_length
)
}
\arguments{
\item{intention_model_bat_speed}{a fitted BRMS intention model for bat speed}

\item{intention_model_swing_length}{a fitted BRMS intention model for swing length}
}
\value{
a table of approach random slopes for bat speed and for swing length
}
\description{
This function extracts only approach random effects from the BRMS intension model because the
full model structure is massive.
}
