% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_primary_fastballs.R
\name{get_primary_fastballs}
\alias{get_primary_fastballs}
\title{Get primary fastball type}
\usage{
get_primary_fastballs(data)
}
\arguments{
\item{data}{a dataset of pitches downloaded by \code{sabRmetrics::download_baseballsavant}}
}
\value{
a table of pitch types, indexed by pitcher_id
}
\description{
This function finds the primary type of fastball for every pitcher in the dataset.
}
