% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_odds.R
\name{log_odds}
\alias{log_odds}
\title{Log-odds

Calculate the log-odds of a probability}
\usage{
log_odds(prob)
}
\arguments{
\item{prob}{a scalar or numeric vector between 0 and 1}
}
\value{
a scalar or numeric vector
}
\description{
Log-odds

Calculate the log-odds of a probability
}
