% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pitch_hit_outcomes.R
\name{predict_pitch_hit_outcomes}
\alias{predict_pitch_hit_outcomes}
\title{Predict pitch outcomes and hit outcomes

This function takes a dataframe of pitch tracking and hit tracking data and appends columns
of predictions from a provided pitch outcome model and hit outcome model.}
\usage{
predict_pitch_hit_outcomes(data, pitch_outcome_model, hit_outcome_model)
}
\arguments{
\item{data}{a dataframe with at least columns \code{balls}, \code{strikes}, \code{is_rhb}, \code{strike_zone_top},
\code{strike_zone_bottom}, \code{x0} and \code{z0} (for pitch outcome predictions) and columns
\code{launch_speed}, \code{launch_angle} and \code{hit_bearing} (for hit oucome predictions)}

\item{pitch_outcome_model}{an object of class "pitch_outcome_model" returned by
\code{predpitchscore::train_pitch_outcome_model}}

\item{hit_outcome_model}{an object of class "xgb.Booster" returned by
\code{predpitchscore::train_hit_outcome_model}}
}
\value{
a dataframe with the same rows and columns as \code{data}, with the following columns added:
prob_swing, prob_hbp, prob_strike, prob_contact, prob_fair, pred_hit,
pitch_value_swing, pitch_value_take, pitch_value hit_pred
}
\description{
Predict pitch outcomes and hit outcomes

This function takes a dataframe of pitch tracking and hit tracking data and appends columns
of predictions from a provided pitch outcome model and hit outcome model.
}
\references{
https://github.com/saberpowers/predictive-pitch-score
}
