% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_squared_up.R
\name{recreate_squared_up}
\alias{recreate_squared_up}
\title{Recreate MLB's "squared-up" metric}
\usage{
recreate_squared_up(data)
}
\arguments{
\item{data}{a dataframe with at least the following columns: \code{ay}, \code{by}, \code{cy}, 'bx', 'bz'
(quadractic coefficients used for calculating plate speed), \code{bat_speed} and \code{launch_speed}.}
}
\value{
a dataframe with the same rows and columns as \code{data}, with the logical column
\code{squared_up} appended.
}
\description{
MLB defines a batted ball as squared up if it is a fair ball and the exit velocity is at least
80\% of (a rough approximation of) its theoretical maximum, given the bat speed and the speed of
the pitch as it crosses the front of home plate. This function calculates the speed of the pitch
at the front of home plate and then tests whether the exit velocity is at least 80\% of its
theoretical max.
}
\references{
http://tangotiger.com/index.php/site/article/statcast-lab-collisions-and-the-perfect-swing
https://x.com/tangotiger/status/1790474439785648493
}
