% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_partial_swings.R
\name{remove_partial_swings}
\alias{remove_partial_swings}
\title{Remove partial swings}
\usage{
remove_partial_swings(swing)
}
\arguments{
\item{swing}{a dataframe with the following coulmns: \code{description}, \code{des}, \code{bat_speed}}
}
\value{
a filtered dataframe after removing rows identified as bunts and checked swings
}
\description{
Filter out bunt attempts and check swings based on bat speed.
}
\examples{
\dontrun{
   data <- sabRmetrics::download_baseballsavant(
     start_date = "2024-07-01",
     end_date = "2024-07-01"
   )
   data_filtered <- data |>
     remove_partial_swings()
}
}
