% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_pitch_outcome.R
\name{summarize_pitch_outcome}
\alias{summarize_pitch_outcome}
\title{Summarize pitch outcome predictions}
\usage{
summarize_pitch_outcome(pred_outcome_pitch)
}
\arguments{
\item{pred_outcome_pitch}{a table with columns \code{prob_swing}, \code{prob_hpb}, \code{prob_strike},
\code{prob_contact}, \code{prob_fair} and \code{pred_hit}}
}
\value{
a table with columns \code{mean_prob_hbp}, \code{mean_prob_ball}, \code{mean_prob_strike},
\code{mean_prob_foul}, \code{mean_prob_fair}, and \code{mean_pred_hit}
}
\description{
Given pre-specified group variables, this function summarizes the average pitch outcome within
each group. Group variables are dropped in the returned table. Whereas pitch outcome predictions
pertain to conditional probabilities (e.g. probability of contact given swing), the summary
probabilities pertain to the multinomial outcome of the pitch (hbp, ball, strike, foul, fair and
expected run value given fair).
}
