% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_intention_model.R
\name{fit_intention_model}
\alias{fit_intention_model}
\title{Fit skew-normal swing intention model}
\usage{
fit_intention_model(swing, swing_metric = c("bat_speed", "swing_length"))
}
\arguments{
\item{swing}{a dataframe of swings}

\item{swing_metric}{the response variable: "bat_speed" or "swing_length"}
}
\value{
a brmsfit object
}
\description{
This function first a skew-normal swing intention model using BRMS.
}
